Basic push works, still needs some ui tuning
