+++ Pleioapp.sln
@@ -5,13 +5,13 @@ VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp", "Pleioapp\Pleioapp.csproj", "{4DDED8A5-F945-46D7-9A45-6B8A2C31A86F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp.iOS", "iOS\Pleioapp.iOS.csproj", "{0051A695-1A2C-4D3A-888C-EBFAF567E421}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp.Droid", "Droid\Pleioapp.Droid.csproj", "{583A438C-C7DB-41E0-AF78-6D374A81425D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp.UITests", "UITests\Pleioapp.UITests.csproj", "{9E34CECF-1BFD-43FD-9D77-292606BBF6F9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp.WP8", "Pleioapp.WP8\Pleioapp.WP8.csproj", "{2D7F7CA0-8919-41DD-B795-A2AC4A459C3F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pleioapp.iOS", "iOS\Pleioapp.iOS.csproj", "{0051A695-1A2C-4D3A-888C-EBFAF567E421}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,24 +53,6 @@ Global
 		{4DDED8A5-F945-46D7-9A45-6B8A2C31A86F}.Release|x64.Build.0 = Release|Any CPU
 		{4DDED8A5-F945-46D7-9A45-6B8A2C31A86F}.Release|x86.ActiveCfg = Release|Any CPU
 		{4DDED8A5-F945-46D7-9A45-6B8A2C31A86F}.Release|x86.Build.0 = Release|Any CPU
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|Any CPU.ActiveCfg = Debug|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|Any CPU.Build.0 = Debug|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|ARM.ActiveCfg = Debug|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhone.ActiveCfg = Debug|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhone.Build.0 = Debug|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhoneSimulator.ActiveCfg = Debug|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhoneSimulator.Build.0 = Debug|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|x64.ActiveCfg = Debug|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|x86.ActiveCfg = Debug|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|Any CPU.ActiveCfg = Release|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|Any CPU.Build.0 = Release|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|ARM.ActiveCfg = Release|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhone.ActiveCfg = Release|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhone.Build.0 = Release|iPhone
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhoneSimulator.ActiveCfg = Release|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhoneSimulator.Build.0 = Release|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|x64.ActiveCfg = Release|iPhoneSimulator
-		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|x86.ActiveCfg = Release|iPhoneSimulator
 		{583A438C-C7DB-41E0-AF78-6D374A81425D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{583A438C-C7DB-41E0-AF78-6D374A81425D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{583A438C-C7DB-41E0-AF78-6D374A81425D}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -161,6 +143,24 @@ Global
 		{2D7F7CA0-8919-41DD-B795-A2AC4A459C3F}.Release|x86.ActiveCfg = Release|x86
 		{2D7F7CA0-8919-41DD-B795-A2AC4A459C3F}.Release|x86.Build.0 = Release|x86
 		{2D7F7CA0-8919-41DD-B795-A2AC4A459C3F}.Release|x86.Deploy.0 = Release|x86
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|Any CPU.ActiveCfg = Debug|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|Any CPU.Build.0 = Debug|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|ARM.ActiveCfg = Debug|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhone.ActiveCfg = Debug|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhone.Build.0 = Debug|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhoneSimulator.ActiveCfg = Debug|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|iPhoneSimulator.Build.0 = Debug|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|x64.ActiveCfg = Debug|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Debug|x86.ActiveCfg = Debug|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|Any CPU.ActiveCfg = Release|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|Any CPU.Build.0 = Release|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|ARM.ActiveCfg = Release|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhone.ActiveCfg = Release|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhone.Build.0 = Release|iPhone
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhoneSimulator.ActiveCfg = Release|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|iPhoneSimulator.Build.0 = Release|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|x64.ActiveCfg = Release|iPhoneSimulator
+		{0051A695-1A2C-4D3A-888C-EBFAF567E421}.Release|x86.ActiveCfg = Release|iPhoneSimulator
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
